sysc: avoid using orpsoc internal classes directly
The problem with using the internal classes directly is
that you have to use the internally generated name,
this in itself is perhaps not such a big issue, the issue
is that the internal name changes when the underlaying verilog
design changes.
This works around this by using the classes through the
top module, which is part of the external api.

git-svn-id: http://opencores.org/ocsvn/openrisc/openrisc/trunk/orpsocv2@862 04c7f051-eaf3-4997-922d-f29d9d8c0f3f